requirements.txt
pandas
numpy
scikit-learn
streamlit
joblib
meteostat
requests
python-dotenv
pytz